removed lexicomn functions from treebankfunctions and put them in a separate module tblex
